Fix LookupTable issues
Fixed small memory leak related to LookupTable
Style
Version 1.5.3
