+++ CMakeLists.txt
@@ -42,7 +42,7 @@ set(MDCM_CMAKE_DIR "${MDCM_SOURCE_DIR}/CMake" CACHE INTERNAL "")
 set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} "${MDCM_CMAKE_DIR}")
 set(MDCM_MAJOR_VERSION 1)
 set(MDCM_MINOR_VERSION 5)
-set(MDCM_BUILD_VERSION 1)
+set(MDCM_BUILD_VERSION 3)
 set(MDCM_VERSION "${MDCM_MAJOR_VERSION}.${MDCM_MINOR_VERSION}.${MDCM_BUILD_VERSION}")
 set(MDCM_API_VERSION "${MDCM_MAJOR_VERSION}.${MDCM_MINOR_VERSION}")
 set(MDCM_LIBRARY_PROPERTIES ${MDCM_LIBRARY_PROPERTIES} VERSION "${MDCM_VERSION}" SOVERSION "${MDCM_API_VERSION}")
