+++ Source/MediaStorageAndFileFormat/mdcmLookupTable.cxx
@@ -22,10 +22,197 @@
 
 #include "mdcmLookupTable.h"
 #include "mdcmTrace.h"
+#include "mdcmSwapper.h"
 #include <set>
 #include <climits>
 #include <cstring>
 #include <cstdint>
+#include <map>
+#include <algorithm>
+#include <iterator>
+#include <vector>
+#include <deque>
+#include <cmath>
+
+// http://blog.goo.ne.jp/satomi_takeo/e/3643e5249b2a9650f9e10ef1c830e8b8
+
+namespace
+{
+
+// abstract class for segment
+template <typename EntryType>
+class Segment
+{
+public:
+  virtual ~Segment() {}
+  typedef std::map<const EntryType *, const Segment *> SegmentMap;
+  virtual bool
+  Expand(const SegmentMap & instances, std::vector<EntryType> & expanded) const = 0;
+  const EntryType *
+  First() const
+  {
+    return first_;
+  }
+  const EntryType *
+  Last() const
+  {
+    return last_;
+  }
+  struct ToMap
+  {
+    std::pair<typename SegmentMap::key_type, typename SegmentMap::mapped_type>
+    operator()(const Segment * segment) const
+    {
+      return std::make_pair(segment->First(), segment);
+    }
+  };
+
+protected:
+  Segment(const EntryType * first, const EntryType * last)
+  {
+    first_ = first;
+    last_ = last;
+  }
+  const EntryType * first_;
+  const EntryType * last_;
+};
+
+// discrete segment (opcode = 0)
+template <typename EntryType>
+class DiscreteSegment : public Segment<EntryType>
+{
+public:
+  typedef typename Segment<EntryType>::SegmentMap SegmentMap;
+  explicit DiscreteSegment(const EntryType * first)
+    : Segment<EntryType>(first, first + 2 + *(first + 1))
+  {}
+  virtual bool
+  Expand(const SegmentMap &, std::vector<EntryType> & expanded) const
+  {
+    std::copy(this->first_ + 2, this->last_, std::back_inserter(expanded));
+    return true;
+  }
+};
+
+// linear segment (opcode = 1)
+template <typename EntryType>
+class LinearSegment : public Segment<EntryType>
+{
+public:
+  typedef typename Segment<EntryType>::SegmentMap SegmentMap;
+  explicit LinearSegment(const EntryType * first)
+    : Segment<EntryType>(first, first + 3)
+  {}
+  virtual bool
+  Expand(const SegmentMap &, std::vector<EntryType> & expanded) const
+  {
+    if (expanded.empty())
+    {
+      // linear segment can't be the first segment.
+      return false;
+    }
+    const EntryType length = *(this->first_ + 1);
+    const EntryType y0 = expanded.back();
+    const EntryType y1 = *(this->first_ + 2);
+    double    y01 = y1 - y0;
+    for (EntryType i = 0; i < length; ++i)
+    {
+      const double v = round(static_cast<double>(y0) + (static_cast<double>(i) / static_cast<double>(length)) * y01);
+      expanded.push_back(static_cast<EntryType>(v));
+    }
+    return true;
+  }
+};
+
+// indirect segment (opcode = 2)
+template <typename EntryType>
+class IndirectSegment : public Segment<EntryType>
+{
+public:
+  typedef typename Segment<EntryType>::SegmentMap SegmentMap;
+  explicit IndirectSegment(const EntryType * first)
+    : Segment<EntryType>(first, first + 2 + 4 / sizeof(EntryType))
+  {}
+  virtual bool
+  Expand(const SegmentMap & instances, std::vector<EntryType> & expanded) const
+  {
+    if (instances.empty())
+    {
+      // some other segments are required as references.
+      return false;
+    }
+    const EntryType *                   first_segment = instances.begin()->first;
+    const unsigned short *              pOffset = reinterpret_cast<const unsigned short *>(this->first_ + 2);
+    unsigned long                       offsetBytes = (*pOffset) | (static_cast<unsigned long>(*(pOffset + 1)) << 16);
+    const EntryType *                   copied_part_head = first_segment + offsetBytes / sizeof(EntryType);
+    typename SegmentMap::const_iterator ppHeadSeg = instances.find(copied_part_head);
+    if (ppHeadSeg == instances.cend())
+    {
+      // referred segment not found
+      return false;
+    }
+    EntryType                           nNumCopies = *(this->first_ + 1);
+    typename SegmentMap::const_iterator ppSeg = ppHeadSeg;
+    while (std::distance(ppHeadSeg, ppSeg) < nNumCopies)
+    {
+      assert(ppSeg != instances.cend());
+      ppSeg->second->Expand(instances, expanded);
+      ++ppSeg;
+    }
+    return true;
+  }
+};
+
+template <typename EntryType>
+void
+ExpandPalette(const EntryType * raw_values, uint32_t length, std::vector<EntryType> & palette)
+{
+  typedef std::deque<Segment<EntryType> *> SegmentList;
+  SegmentList                              segments;
+  const EntryType *                        raw_seg = raw_values;
+  while ((std::distance(raw_values, raw_seg) * sizeof(EntryType)) < length)
+  {
+    Segment<EntryType> * segment = nullptr;
+    if (*raw_seg == 0)
+    {
+      segment = new DiscreteSegment<EntryType>(raw_seg);
+    }
+    else if (*raw_seg == 1)
+    {
+      segment = new LinearSegment<EntryType>(raw_seg);
+    }
+    else if (*raw_seg == 2)
+    {
+      segment = new IndirectSegment<EntryType>(raw_seg);
+    }
+    if (segment)
+    {
+      segments.push_back(segment);
+      raw_seg = segment->Last();
+    }
+    else
+    {
+      // invalid opcode
+      break;
+    }
+  }
+  typename Segment<EntryType>::SegmentMap instances;
+  std::transform(
+    segments.begin(), segments.end(), std::inserter(instances, instances.end()), typename Segment<EntryType>::ToMap());
+  typename SegmentList::iterator ppSeg = segments.begin();
+  typename SegmentList::iterator endOfSegments = segments.end();
+  for (; ppSeg != endOfSegments; ++ppSeg)
+  {
+    (*ppSeg)->Expand(instances, palette);
+  }
+  ppSeg = segments.begin();
+  for (; ppSeg != endOfSegments; ++ppSeg)
+  {
+    delete *ppSeg;
+  }
+}
+
+}
 
 namespace mdcm
 {
@@ -33,8 +220,8 @@ namespace mdcm
 bool
 LookupTable::Initialized() const
 {
-  bool b1 = BitSample != 0;
-  bool b2 = Internal.BitSize[0] != 0 && Internal.BitSize[1] != 0 && Internal.BitSize[2] != 0;
+  const bool b1 = BitSample != 0;
+  const bool b2 = Internal.BitSize[0] != 0 && Internal.BitSize[1] != 0 && Internal.BitSize[2] != 0;
   return b1 && b2;
 }
 
@@ -79,11 +266,15 @@ LookupTable::InitializeLUT(LookupTableType type,
                            unsigned short  subscript,
                            unsigned short  bitsize)
 {
+  if (!(static_cast<int>(type) >= 0 && static_cast<int>(type) <= 2))
+  {
+    mdcmWarningMacro("Error in InitializeLUT,  LookupTableType " << type);
+	return;
+  }
   if (bitsize != 8 && bitsize != 16)
   {
     return;
   }
-  assert(bitsize == 8 || bitsize == 16);
   if (length == 0)
   {
     Internal.Length[type] = 65536;
@@ -124,9 +315,14 @@ LookupTable::SetLUT(LookupTableType enhanced_type, const unsigned char * array, 
   {
     type = BLUE;
   }
-  if (!Internal.Length[type])
+  if (!(static_cast<int>(type) >= 0 && static_cast<int>(type) <= 2))
   {
-    mdcmDebugMacro("Need to set length first");
+    mdcmWarningMacro("Error in SetLUT,  LookupTableType " << enhanced_type);
+	return;
+  }
+  if (Internal.Length[type] == 0)
+  {
+    mdcmWarningMacro("Error in SetLUT, Length[" << type << "] is 0");
     return;
   }
   if (!IncompleteLUT)
@@ -155,7 +351,7 @@ LookupTable::SetLUT(LookupTableType enhanced_type, const unsigned char * array, 
     }
     else
     {
-      unsigned int offset = 0;
+      const unsigned int offset = 0;
       assert(mult2 == 2);
       for (unsigned int i = 0; i < Internal.Length[type]; ++i)
       {
@@ -182,8 +378,65 @@ LookupTable::SetLUT(LookupTableType enhanced_type, const unsigned char * array, 
 }
 
 void
+LookupTable::SetSegmentedLUT(LookupTableType enhanced_type, const unsigned char * array, unsigned int length)
+{
+  // Supplemental SUPPLRED = 4
+  LookupTableType type = enhanced_type;
+  if (type == SUPPLRED)
+  {
+    type = RED;
+  }
+  else if (type == SUPPLGREEN)
+  {
+    type = GREEN;
+  }
+  else if (type == SUPPLBLUE)
+  {
+    type = BLUE;
+  }
+  if (!(static_cast<int>(type) >= 0 && static_cast<int>(type) <= 2))
+  {
+    mdcmWarningMacro("Error in SetSegmentedLUT, LookupTableType " << enhanced_type);
+	return;
+  }
+  if (Internal.Length[type] == 0)
+  {
+    mdcmWarningMacro("Error in SetSegmentedLUT, Length[" << type << "] is 0");
+    return;
+  }
+  if (BitSample == 8)
+  {
+    assert(0); // TODO
+  }
+  else if (BitSample == 16)
+  {
+    assert(length % 2 == 0);
+    unsigned char * copy = new unsigned char[length];
+    for (size_t x = 0; x < length; ++x)
+    {
+      copy[x] = array[x];
+    }
+    void *                varray = static_cast<void*>(copy);
+    uint16_t *            array16 = static_cast<uint16_t *>(varray);
+    std::vector<uint16_t> palette;
+    unsigned int          num_entries = GetLUTLength(type);
+    palette.reserve(num_entries);
+    SwapperNoOp::SwapArray(array16, length / 2);
+    ExpandPalette(array16, length, palette);
+    const void * vpalette = static_cast<const void*>(palette.data());
+    SetLUT(enhanced_type, static_cast<const unsigned char *>(vpalette), static_cast<unsigned int>(palette.size() * 2));
+    delete [] copy;
+  }
+}
+
+void
 LookupTable::GetLUT(LookupTableType type, unsigned char * array, unsigned int & length) const
 {
+  if (!(static_cast<int>(type) >= 0 && static_cast<int>(type) <= 2))
+  {
+    mdcmWarningMacro("Error in GetLUT, LookupTableType " << type);
+	return;
+  }
   if (BitSample == 8)
   {
     const unsigned int mult = Internal.BitSize[type] / 8;
@@ -222,7 +475,11 @@ LookupTable::GetLUTDescriptor(LookupTableType  type,
                               unsigned short & subscript,
                               unsigned short & bitsize) const
 {
-  assert(type >= RED && type <= BLUE);
+  if (!(static_cast<int>(type) >= 0 && static_cast<int>(type) <= 2))
+  {
+    mdcmWarningMacro("Error in GetLUTDescriptor, LookupTableType " << type);
+	return;
+  }
   if (Internal.Length[type] == 65536)
   {
     length = 0;
@@ -271,14 +528,6 @@ void
 LookupTable::SetBlueLUT(const unsigned char * blue, unsigned int length)
 {
   SetLUT(BLUE, blue, length);
-}
-
-inline void
-printrgb(const unsigned char * rgb)
-{
-  std::cout << static_cast<int>(rgb[0]) << ','
-            << static_cast<int>(rgb[1]) << ','
-            << static_cast<int>(rgb[2]);
 }
 
 void
@@ -567,9 +816,5 @@ LookupTable::GetBitSample() const
 {
   return BitSample;
 }
-
-void
-LookupTable::Print(std::ostream &) const
-{}
 
 } // end namespace mdcm
