+++ Source/MediaStorageAndFileFormat/mdcmLookupTable.h
@@ -24,9 +24,8 @@
 #define MDCMLOOKUPTABLE_H
 
 #include "mdcmTypes.h"
-#include "mdcmObject.h"
-#include <cstdlib>
 #include <vector>
+#include <istream>
 
 namespace mdcm
 {
@@ -43,7 +42,7 @@ public:
 /**
  * LookupTable class
  */
-class MDCM_EXPORT LookupTable : public Object
+class MDCM_EXPORT LookupTable
 {
 public:
   typedef enum
@@ -58,8 +57,6 @@ public:
     UNKNOWN
   } LookupTableType;
 
-  LookupTable() = default;
-  virtual ~LookupTable() = default;
   bool
   Initialized() const;
   void
@@ -68,8 +65,10 @@ public:
   Allocate(unsigned short = 8 /*bitsample*/);
   void InitializeLUT(LookupTableType, unsigned short, unsigned short, unsigned short);
   unsigned int GetLUTLength(LookupTableType) const;
-  virtual void
+  void
   SetLUT(LookupTableType, const unsigned char *, unsigned int);
+  void
+  SetSegmentedLUT(LookupTableType, const unsigned char *, unsigned int);
   void
   GetLUT(LookupTableType, unsigned char *, unsigned int &) const;
   void
@@ -100,8 +99,6 @@ public:
   WriteBufferAsRGBA(const unsigned char *);
   unsigned short
   GetBitSample() const;
-  void
-  Print(std::ostream &) const override;
 
 protected:
   LookupTableInternal Internal{};
